requirements.txt
speechrecognition
pyaudio
gtts
openai
playsound==1.2.2
python-dotenv
flask-cors
pydub
